Fix truncate impl in JS bindings

+++ core/storage/database.rs
@@ -18,7 +18,7 @@ pub trait DatabaseStorage: Send + Sync {
     ) -> Result<Completion>;
     fn sync(&self, c: Completion) -> Result<Completion>;
     fn size(&self) -> Result<u64>;
-    fn truncate(&self, len: u64, c: Completion) -> Result<()>;
+    fn truncate(&self, len: usize, c: Completion) -> Result<()>;
 }
 
 #[cfg(feature = "fs")]
@@ -72,7 +72,7 @@ impl DatabaseStorage for DatabaseFile {
     }
 
     #[instrument(skip_all, level = Level::INFO)]
-    fn truncate(&self, len: u64, c: Completion) -> Result<()> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         self.file.truncate(len, c)?;
         Ok(())
     }
@@ -131,7 +131,7 @@ impl DatabaseStorage for FileMemoryStorage {
     }
 
     #[instrument(skip_all, level = Level::INFO)]
-    fn truncate(&self, len: u64, c: Completion) -> Result<()> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         let _ = self.file.truncate(len, c)?;
         Ok(())
     }
